Fixed crashes on ze_atix_panic_v1
Explanation: Originally i thought the prop crashes because it spawns with solid 6 without a .phy model and that my fix is somehow not catching that. well i was wrong. the map itself breaks it, my plugin doesnt catch it because the model actually DOES have a proper .phy model, but the map sets that prop to solid 0 AFTER the prop already spawned (model already in solidity list as solid)